+++ colls/gather/gather.c
@@ -1,6 +1,7 @@
 #include "mpi.h"
 #include "util.h"
 
+#include <stddef.h>
 #include <string.h>
 
 int MPI_Gather_basic(const void *sendbuf, int sendcount, MPI_Datatype sendtype, void *recvbuf,
@@ -12,10 +13,10 @@ int MPI_Gather_basic(const void *sendbuf, int sendcount, MPI_Datatype sendtype, 
     if (rank == root) {
         for (int i = 0; i < size; i++) {
             if (i != root) {
-                MPI_Recv((char *)recvbuf + i * recvcount * nanompi_get_dtype_size(recvtype),
+                MPI_Recv((char *)recvbuf + ((ptrdiff_t)(i * recvcount * nanompi_get_dtype_size(recvtype))),
                          recvcount, recvtype, i, 0, comm, MPI_STATUS_IGNORE);
             } else {
-                memcpy((char *)recvbuf + i * recvcount * nanompi_get_dtype_size(recvtype), sendbuf,
+                memcpy((char *)recvbuf + ((ptrdiff_t)(i * recvcount * nanompi_get_dtype_size(recvtype))), sendbuf,
                        sendcount * nanompi_get_dtype_size(sendtype));
             }
         }
